+++ linux/include/custom_share/custom_share_plugin.h
@@ -13,13 +13,13 @@ G_BEGIN_DECLS
 
 typedef struct _CustomSharePlugin CustomSharePlugin;
 typedef struct {
-  GObjectClass parent_class;
+    GObjectClass parent_class;
 } CustomSharePluginClass;
 
 FLUTTER_PLUGIN_EXPORT GType custom_share_plugin_get_type();
 
 FLUTTER_PLUGIN_EXPORT void custom_share_plugin_register_with_registrar(
-    FlPluginRegistrar* registrar);
+        FlPluginRegistrar* registrar);
 
 G_END_DECLS