+++ test/custom_share_method_channel_test.dart
@@ -1,9 +1,8 @@
 /// Tests for the [MethodChannelCustomShare] implementation of the custom_share plugin.
 ///
 /// This file contains unit tests to verify the behavior of [MethodChannelCustomShare]
-/// methods [shareText] and [shareFiles] using a mocked [MethodChannel].
+/// methods [shareText] and [shareFile] using a mocked [MethodChannel].
 library;
-
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:custom_share/custom_share_method_channel.dart';
@@ -12,15 +11,11 @@ import 'package:custom_share/custom_share_method_channel.dart';
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
   const MethodChannel channel = MethodChannel('custom_share');
-  MethodChannelCustomShare platform = MethodChannelCustomShare();
 
   setUp(() {
     /// Sets up a mock method call handler for the [channel].
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(
-          channel,
-          (MethodCall methodCall) async => 'success',
-        );
+        .setMockMethodCallHandler(channel, (MethodCall methodCall) async => 'success');
   });
 
   tearDown(() {
@@ -31,11 +26,17 @@ void main() {
 
   test('shareText', () async {
     /// Tests that [shareText] returns 'success' for a valid text input.
-    expect(await platform.shareText(text: 'test'), 'success');
+    expect(await MethodChannelCustomShare.shareText(text: 'test'), 'success');
   });
 
-  test('shareFiles', () async {
-    /// Tests that [shareFiles] returns 'success' for a valid file path list.
-    expect(await platform.shareFiles(filePaths: ['/path/to/file']), 'success');
+  test('shareFile', () async {
+    /// Tests that [shareFile] returns 'success' for a valid file path and MIME type.
+    expect(
+      await MethodChannelCustomShare.shareFile(
+        filePath: '/path/to/file.png',
+        mimeType: 'image/png',
+      ),
+      'success',
+    );
   });
 }